==> notebooks/01_EDA_artists.ipynb <==
# Data Exploration – MoMA Art Dataset

---

## Data Loading

import sys
import os

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import importlib
import src.data_loader as dl

importlib.reload(dl)

artists = dl.load_artists()


## 2. General Structure
- Dataset size
- Columns and data types
- First few rows

artists.info()

artists.head()


## 3. Missing Values and Data Quality

artists.isna().sum()

artists.duplicated().sum()

## 4. Distribution and Cardinality
- Categorical columns
- Numerical columns: ID, beginDate, endDate, ULAN
- Cardinality (how many unique values?)

## 5. Dataset Relationships
- Relationships between `Artworks` and `Artists`: only artist name
- Possible primary/foreign keys
