==> notebooks/02_cleaning_artists.ipynb <==
import sys
import os

import numpy as np

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import importlib
import src.data_loader as dl

importlib.reload(dl)
artists = dl.load_artists()

artists['Nationality'] = artists['Nationality'].fillna('Unknown')
artists['Gender'] = artists['Gender'].fillna('Unknown')

artists['Gender'] = artists['Gender'].str.strip().str.lower()
artists['Nationality'] = artists['Nationality'].str.strip().str.title()

artists['EndDate'] = artists['EndDate'].replace(0, np.nan)

artists.to_csv("../outputs/artists_clean.csv", index=False)

artists.info()
artists.head()
artists['Gender'].value_counts()
